pyproject.toml
[tool.poetry]
name = "agent-tracking"
version = "0.1.0"
description = "A software project for agent tracking to disaply all changement done by agent"
authors = ["tom.houee@hymaia.com"]
readme = "README.md"
packages = [{include = "agent_tracking", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# dependencies for code analysis, UML generation and visualization
pandas = "^2.1"
matplotlib = "^3.8"
seaborn = "^0.13"

# we keep builtins to avoid heavy extras, XML is constructed manually
gitpython = "^3.1.46"
radon = "^6.0.1"
pyvis = "^0.3.2"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
python-dotenv = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.1.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
agent-tracking = "agent_tracking.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
